+++ src/sagas/GetProfile.test.ts
@@ -4,7 +4,7 @@ import { delay } from 'redux-saga'
 import { takeEvery, takeLatest, call, put } from 'redux-saga/effects'
 
 import { Actions } from '../actions/index'
-import { GetProfileFrost, GetProfileSaga, GetProfile } from './GetProfile.saga'
+import { GetProfileFrost, GetProfileSaga, GetProfile } from './GetProfile'
 
 describe('SignInSaga()', async (should: any) => {
   const { assert } = should()
