+++ tests/unit/index.ts
@@ -3,9 +3,9 @@
 import '../../src/components/atoms/Button/Button.test'
 import '../../src/components/atoms/Input/Input.test'
 import '../../src/components/molecules/Form/Form.test'
-import '../../src/reducers/LoadingPage.reducer.test'
-import '../../src/reducers/Profile.reducer.test'
-import '../../src/reducers/SignIn.reducer.test'
-import '../../src/reducers/User.reducer.test'
-import '../../src/sagas/GetProfile.saga.test'
-import '../../src/sagas/SignIn.saga.test'
+import '../../src/reducers/LoadingPage.test'
+import '../../src/reducers/Profile.test'
+import '../../src/reducers/SignIn.test'
+import '../../src/reducers/User.test'
+import '../../src/sagas/GetProfile.test'
+import '../../src/sagas/SignIn.test'
